Extract text-related functions to new module, TextReader
